Check mount options of tmpfs and devtmpfs filesystems
As a coarse recommendation, sum of sizes of all tmpfs and devtmpfs
mounts should be less than the size of the RAM to prevent out of
memory situations. Without the size= mount option, default size of
each mount is 50% of RAM. Add a simple check for presence of size=
options.

Check also for presence of nr_inodes= options, since memory accounting
in current tmpfs implementation does not cover all use cases and
memory usage can be limited with number of inodes too. The default
without nr_inodes= is half of the number of RAM pages on modern
hardware, so this should also be lowered relative to size option.

This check could be extended for BSDs since their tmpfs
implementations also have similar size= and inodes= options.

Signed-off-by: Topi Miettinen <toiwoton@gmail.com>