Use nawk and GNU echo on Solaris
The default Solaris awk does not support the commands used
by Lynis. As osdetection uses awk, too, nawk is used there on
Solaris as well.

The echo command on Solaris does not support -n.
Lynis tries to use /usr/ucb/echo, but it is not present on Solaris 11.
Thus GNU echo is used or printf.