Add support for Solaris IPS
IPS (Image Packaging System) is the new package system of Solaris 11.
This commit adds support for it.

New tests PKGS-7338 to -7341 detect this package manager, search for
updates and check if the support repository and patch metadata are
installed.

Additionally `pkg audit` is not detected as audit tool any more.