Find sshd on Solaris and simplify version check
Solaris has sshd at the location /usr/lib/ssh/sshd instead of the common
binary locations.

The OpenSSH version number could not have been gathered with nawk, as
it does not support comparison with ~. Without a valid version number,
the configuration check was skipped. This commit fixes it, thus the sshd
configuration is now (again) checked on Solaris.

See: 3987911